Two changes: 1. Removed '.' from import statements for local modules 2. Set load_kb docstring to raw docstring 3. Set unused classes to private
1. Removed '.' from import statements as this
was causing sphinx.ext.autodoc to fail as it
tries to find a parent module named ''. It,
of course, doesn't find the module and just
gives up on generating the docs for this module.

2. Set load_kb's docstring to a raw docstring,
so that it renders backslashes properly. Also
made minor amendments to clarify the example.

3. Set unused classes to private by adding the '_' prefix.
This leverages sphinx.ext.autodoc's default behaviour i.e.
it ignores all private classes. See /doc/README.md
for more information on this.